requirements.txt
# HTTP请求
requests>=2.31.0
urllib3>=2.0.0

# 环境配置
python-dotenv>=1.0.0

# HTML和Markdown处理
beautifulsoup4>=4.12.0
markdown>=3.5.0
lxml>=4.9.0
html5lib>=1.1

# 图片处理
Pillow>=10.0.0

# API客户端
openai>=1.0.0
dashscope>=1.13.0

# 工具库
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.65.0 